try to suppress warnings by explicitly setting anndata dtype in tests

+++ cell_type_mapping_tree/tests/validation/test_validation_utils.py
@@ -94,7 +94,7 @@ def create_h5ad_file(
             f"do not know what density={density} means")
 
     a_data = anndata.AnnData(
-        X=data, obs=obs, var=var)
+        X=data, obs=obs, var=var, dtype=data.dtype)
 
     a_data.write_h5ad(output_path)
 
@@ -364,7 +364,7 @@ def test_is_x_integers_layers(tmp_dir_fixture, is_sparse, is_int):
     if is_sparse:
         layer = scipy_sparse.csr_matrix(layer)
 
-    a_data = anndata.AnnData(X=x, layers={'garbage': layer})
+    a_data = anndata.AnnData(X=x, layers={'garbage': layer}, dtype=x.dtype)
 
     h5ad_path = mkstemp_clean(
         dir=tmp_dir_fixture,
@@ -406,7 +406,7 @@ def test_get_minmax_integers_layers(tmp_dir_fixture, is_sparse):
     if is_sparse:
         layer = scipy_sparse.csr_matrix(layer)
 
-    a_data = anndata.AnnData(X=x, layers={'garbage': layer})
+    a_data = anndata.AnnData(X=x, layers={'garbage': layer}, dtype=x.dtype)
 
     h5ad_path = mkstemp_clean(
         dir=tmp_dir_fixture,
